+++ src/stockpyl/ssm_serial.py
@@ -943,8 +943,7 @@ def _preprocess_parameters(num_nodes=None, node_order_in_system=None, node_order
 	# Validate more parameters.
 	if any(c is None for c in echelon_holding_cost_dict.values()): raise ValueError("echelon_holding_cost cannot be None for any node")
 	if stockout_cost < 0: raise ValueError("stockout_cost must be non-negative")
-#	if any(L is None for L in lead_time_dict.values()): raise ValueError("lead_time cannot be None for any node")
-	if not all(lead_time_dict.values()): raise ValueError("lead_time cannot be None for any node")
+	if any(L is None for L in lead_time_dict.values()): raise ValueError("lead_time cannot be None for any node")
 	if any(l < 0 for l in lead_time_dict.values()): raise ValueError("lead_time must be non-negative for every node")
 	
 	return old_to_new_dict, num_nodes, echelon_holding_cost_dict, lead_time_dict, stockout_cost, demand_source