+++ tests/test_ssm_serial.py
@@ -451,6 +451,31 @@ class TestOptimizeBaseStockLevels(unittest.TestCase):
 				x_num=100, d_num=10
 			)
 
+	def test_zero_leadtime(self):
+		"""Test that ssm_serial.optimize_base_stock_levels() works when LT = 0. 
+		(Issue 173 reported that this produced an error.)
+		"""
+		print_status('TestIssue173', 'test_multisource_orders()')
+
+		# Build network.
+		network = serial_system(
+			num_nodes=2,
+			node_order_in_system=[2, 1],
+			echelon_holding_cost=[1, 1],
+			local_holding_cost=[1, 2],
+			shipment_lead_time=[0, 0],
+			stockout_cost=2,
+			demand_type='N',
+			mean=50,
+			standard_deviation=10
+		)
+		# Optimize echelon base-stock levels.
+		S_star, C_star = optimize_base_stock_levels(network=network)
+		print(f"Optimal echelon base-stock levels = {S_star}")
+		print(f"Optimal expected cost per period = {C_star}")
+
+		self.assertDictEqual(S_star, {1: 0, 2:0})
+		self.assertEqual(C_star, 0)
 
 class TestNewsvendorHeuristic(unittest.TestCase):